% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workshop-deps.R
\name{admin_update_deps}
\alias{admin_update_deps}
\title{Update r3's dependencies from the workshops' \code{DESCRIPTION} files}
\usage{
admin_update_deps()
}
\value{
Doesn't output anything, used for the side effect to update the
\code{DESCRIPTION} file.
}
\description{
Compares with all the workshops
}
