% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_git_config}
\alias{check_git_config}
\title{Checks your computer's Git config settings.}
\usage{
check_git_config()
}
\value{
Prints config settings to screen.
}
\description{
Checks your computer's Git config settings.
}
