% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_setup}
\alias{check_setup}
\title{Check your setup in preparation for the r-cubed course.}
\usage{
check_setup()
}
\value{
Don't output anything, only runs checks on the setup.
}
\description{
Check your setup in preparation for the r-cubed course.
}
