% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_system}
\alias{check_system}
\alias{check_r_version}
\alias{check_rstudio_version}
\alias{check_git_config}
\alias{check_setup}
\title{Runs checks on your system for R, RStudio, and Git.}
\usage{
check_r_version()

check_rstudio_version()

check_git_config()

check_setup()
}
\value{
Prints information on given check.
}
\description{
Runs checks on your system for R, RStudio, and Git.
}
\section{Functions}{
\itemize{
\item \code{check_r_version}: Check version of R that is installed.

\item \code{check_rstudio_version}: Check version of RStudio that is installed.

\item \code{check_git_config}: Checks your computer's Git config settings.

\item \code{check_setup}: Check your setup in preparation for the r-cubed course.
}}

