% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-rmd.R
\name{create_doc}
\alias{create_doc}
\alias{create_qmd_doc}
\alias{create_rmd_doc}
\title{Create document in the doc folder.}
\usage{
create_qmd_doc()

create_rmd_doc()
}
\value{
Creates a Markdown document in the \verb{doc/} folder.
}
\description{
Make a draft of either an R Markdown document or a Quarto document
}
\section{Functions}{
\itemize{
\item \code{create_qmd_doc()}: Creates a simple draft of a Quarto document.

\item \code{create_rmd_doc()}: Creates a simple draft of a R Markdown document.

}}
