% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-doc.R
\name{create_qmd}
\alias{create_qmd}
\alias{create_learning_qmd}
\alias{create_cleaning_qmd}
\alias{create_sandbox_qmd}
\title{Create document in the \verb{docs/} folder.}
\usage{
create_learning_qmd()

create_cleaning_qmd()

create_sandbox_qmd()
}
\value{
Creates a Quarto Markdown document in the \verb{docs/} folder.
}
\description{
Make a draft of a Quarto document to use for the workshops.
}
\section{Functions}{
\itemize{
\item \code{create_learning_qmd()}: Create a \code{learning.qmd} file in the \verb{docs/} folder.

\item \code{create_cleaning_qmd()}: Create a \code{cleaning.qmd} file in the \verb{docs/} folder.

\item \code{create_sandbox_qmd()}: Create a \code{sandbox.qmd} file in the \verb{docs/} folder.

}}
