% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-doc.R
\name{create_qmd_doc}
\alias{create_qmd_doc}
\title{Create document in the \verb{docs/} folder.}
\usage{
create_qmd_doc()
}
\value{
Creates a Quarto Markdown document in the \verb{docs/} folder.
}
\description{
Make a draft of a Quarto document to use for the workshops.
}
