% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-rmd.R
\name{create_rmarkdown_doc}
\alias{create_rmarkdown_doc}
\title{Create a basic R Markdown document in the doc folder.}
\usage{
create_rmarkdown_doc()
}
\value{
Creates an R Markdown file called \code{lesson.Rmd} in the \verb{doc/} folder.
}
\description{
Create a basic R Markdown document in the doc folder.
}
