% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-helpers.R
\name{fix_git_remotes}
\alias{fix_git_remotes}
\title{Fix some Git remote issues that occur during the course.}
\usage{
fix_git_remotes()
}
\value{
Used for the side effect of connecting to Git. Returns NULL invisibly.
}
\description{
Fix some Git remote issues that occur during the course.
}
