% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-versions.R
\name{get_allowed_r_versions}
\alias{get_allowed_r_versions}
\title{Get a list of the allowed R versions for the rostools workshops.}
\usage{
get_allowed_r_versions()
}
\value{
A character vector of allowed R versions.
}
\description{
The allowed versions are those that are at maximum a year old from today,
sorted by the oldest first.
}
\examples{
get_allowed_r_versions()
}
