% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-versions.R
\name{get_allowed_rstudio_versions}
\alias{get_allowed_rstudio_versions}
\title{Get a list of the allowed RStudio versions for the rostools workshops.}
\usage{
get_allowed_rstudio_versions()
}
\value{
A character vector of allowed RStudio versions.
}
\description{
The versions are those that are at a maximum one year old from today,
sorted by the oldest first.
}
\examples{
get_allowed_rstudio_versions()
}
