% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-packages.R
\name{install_packages}
\alias{install_packages}
\alias{install_packages_intermediate}
\alias{install_packages_advanced}
\alias{install_packages_introduction}
\title{Install the R packages necessary for the specific R workshop.}
\usage{
install_packages_intermediate()

install_packages_advanced()

install_packages_introduction()
}
\value{
Outputs nothing, used to install packages for the R workshop.
}
\description{
Install the R packages necessary for the specific R workshop.
}
\section{Functions}{
\itemize{
\item \code{install_packages_intermediate()}: Install necessary packages for the intermediate workshop.

\item \code{install_packages_advanced()}: Install necessary packages for the advanced workshop.

\item \code{install_packages_introduction()}: Install necessary packages for the introduction workshop.

}}
