% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-packages.R
\name{install_packages_advanced}
\alias{install_packages_advanced}
\title{Install the R packages necessary for the Advanced R course.}
\usage{
install_packages_advanced()
}
\value{
Outputs nothing, used to install packages for the R course.
}
\description{
Install the R packages necessary for the Advanced R course.
}
