% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{mmash_data_link}
\alias{mmash_data_link}
\title{Link to MMASH zip data file, just in case.}
\format{
An object of class \code{character} of length 1.
}
\usage{
mmash_data_link
}
\description{
Link to MMASH zip data file, just in case.
}
\keyword{datasets}
