% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-links.R
\name{open_feedback_survey}
\alias{open_feedback_survey}
\title{Opens daily feedback survey for the workshops.}
\usage{
open_feedback_survey()
}
\value{
Nothing, used for the side effect of opening the survey link.
}
\description{
Opens daily feedback survey for the workshops.
}
