% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-links.R
\name{open_url}
\alias{open_url}
\alias{open_github_join}
\alias{open_feedback_survey}
\alias{open_feedback_survey_intro}
\alias{open_feedback_survey_intermediate}
\alias{open_feedback_survey_advanced}
\title{Opens a browser to course resources.}
\usage{
open_github_join()

open_feedback_survey()

open_feedback_survey_intro()

open_feedback_survey_intermediate()

open_feedback_survey_advanced()
}
\description{
Opens a browser to course resources.
}
\section{Functions}{
\itemize{
\item \code{open_github_join()}: Opens GitHub account creation website.

\item \code{open_feedback_survey()}: Opens daily feedback survey for all courses.

\item \code{open_feedback_survey_intro()}: Opens daily feedback survey for intro course.

\item \code{open_feedback_survey_intermediate()}: Opens daily feedback survey for intermediate course.

\item \code{open_feedback_survey_advanced()}: Opens daily feedback survey for advanced course.

}}
