% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3-package.R
\docType{package}
\name{r3-package}
\alias{r3}
\alias{r3-package}
\title{r3: Companion Package for Rostools R Workshops}
\description{
Contains the helper functions and other files to make it easier for learners during (and after) the workshop.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rostools/r3}
  \item Report bugs at \url{https://github.com/rostools/r3/issues}
}

}
\author{
\strong{Maintainer}: Luke Johnston \email{lwjohnst@gmail.com} (\href{https://orcid.org/0000-0003-4169-2616}{ORCID})

}
\keyword{internal}
