% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_git_config}
\alias{setup_git_config}
\title{Set the user's git configuration.}
\usage{
setup_git_config()
}
\value{
Modifies global git configuration.
}
\description{
Set the user's git configuration.
}
