% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_git_settings}
\alias{setup_git_settings}
\title{Set some Git config settings like main as default and rebase as false.}
\usage{
setup_git_settings()
}
\value{
Invisible, values of previous settings.
}
\description{
Set some Git config settings like main as default and rebase as false.
}
