% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_rproj}
\alias{setup_rproj}
\title{Setup the R project file to simplify learning in the course.}
\usage{
setup_rproj()
}
\value{
Invisibly outputs the contents of the \code{.Rproj} file. Writes to file
as side effect.
}
\description{
Setup the R project file to simplify learning in the course.
}
