% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_rstudio}
\alias{setup_rstudio}
\title{Set some RStudio settings that help make things smoother during the course.}
\usage{
setup_rstudio()
}
\value{
Invisibly returns NULL, used for side effect of setting options in RStudio.
}
\description{
Set some RStudio settings that help make things smoother during the course.
}
