% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3-package.R
\docType{data}
\name{useful_learning_sites_list}
\alias{useful_learning_sites_list}
\title{List of some useful sites for learning R.}
\format{
A column with 8 rows and 1 variables:
\describe{
\item{Site}{List of learning sites, as a Markdown link.}
}
}
\usage{
useful_learning_sites_list
}
\description{
A one-column dataset containing a Markdown formatted list of learning sites
with URLs.
}
\keyword{datasets}
