% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3-package.R
\docType{data}
\name{useful_packages_list}
\alias{useful_packages_list}
\title{List of some useful packages to use in R.}
\format{
A column with 27 rows and 3 variables:
\describe{
\item{Package}{List of useful packages, as a Markdown link.}
\item{Title}{Package title.}
\item{Description}{Package description.}
}
}
\usage{
useful_packages_list
}
\description{
A data frame containing a Markdown formatted list of packages
with URLs, their titles, and their description.
}
\keyword{datasets}
