% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-packages.R
\name{install_packages}
\alias{install_packages}
\alias{install_packages_advanced}
\alias{install_packages_introduction}
\title{Install the R packages necessary for the Advanced R course.}
\usage{
install_packages_advanced()

install_packages_introduction()
}
\value{
Outputs nothing, used to install packages for the R course.
}
\description{
Install the R packages necessary for the Advanced R course.
}
\section{Functions}{
\itemize{
\item \code{install_packages_advanced()}: Install necessary packages for the advanced course.

\item \code{install_packages_introduction()}: Install necessary packages for the introduction course.

}}
