% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-links.R
\name{open_url}
\alias{open_url}
\alias{open_course_material}
\alias{open_course_website}
\alias{open_schedule}
\alias{open_github}
\alias{open_pre_survey}
\alias{open_proj_management}
\alias{open_version_control}
\alias{open_data_wrangling}
\alias{open_data_visualization}
\alias{open_reproducible_documents}
\alias{open_code_of_conduct}
\alias{open_syllabus}
\alias{open_assignment}
\title{Opens a browser to course resources.}
\usage{
open_course_material()

open_course_website()

open_schedule()

open_github()

open_pre_survey()

open_proj_management()

open_version_control()

open_data_wrangling()

open_data_visualization()

open_reproducible_documents()

open_code_of_conduct()

open_syllabus()

open_assignment()
}
\description{
Opens a browser to course resources.
}
\section{Functions}{
\itemize{
\item \code{open_course_material}: Opens website for course material.

\item \code{open_course_website}: Opens website for DDA R course, March 2020.

\item \code{open_schedule}: Opens schedule for the DDA R course.

\item \code{open_github}: Opens GitHub account creation website.

\item \code{open_pre_survey}: Opens pre-course survey.

\item \code{open_proj_management}: Function to take you right to Project Management session.

\item \code{open_version_control}: Function to take you right to Version Control session.

\item \code{open_data_wrangling}: Function to take you right to Data Management and Wrangling session.

\item \code{open_data_visualization}: Function to take you right to Data Visualization session.

\item \code{open_reproducible_documents}: Function to take you right to Reproducible Documents session.

\item \code{open_code_of_conduct}: Opens the Code of Conduct webpage.

\item \code{open_syllabus}: Function to take you right to the Syllabus webpage.

\item \code{open_assignment}: Function to take you to the group project.
}}

