% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3-package.R
\docType{package}
\name{r3-package}
\alias{r3}
\alias{r3-package}
\title{r3: Companion Package for the R3 Introduction to R
    Course.}
\description{
Contains the helper functions and other files to
    make it easier for learners during (and after) the course.
}
\author{
\strong{Maintainer}: Luke Johnston \email{lwjohnst@gmail.com} (\href{https://orcid.org/0000-0003-4169-2616}{ORCID})

}
\keyword{internal}
